requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kucoin-futures-python==1.0.6
kucoin-python==1.0.11
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
pandas-ta==0.3.14b0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.7
typing-extensions==4.5.0
urllib3==1.26.15
websockets==10.4
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
